+++ src/main/java/codes/towel/RCCAssistance/command/status/StatusCommandsExecutor.java
@@ -1,8 +1,8 @@
-package codes.flappy.RCCAssistance.command.status;
+package codes.towel.RCCAssistance.command.status;
 
-import codes.flappy.RCCAssistance.ResponseEmbedBuilder;
-import codes.flappy.RCCAssistance.command.CommandExecutor;
-import codes.flappy.RCCAssistance.command.UnknownCommandException;
+import codes.towel.RCCAssistance.ResponseEmbedBuilder;
+import codes.towel.RCCAssistance.command.CommandExecutor;
+import codes.towel.RCCAssistance.command.UnknownCommandException;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.entities.Activity;
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
